[APG-454] refactor
